Fix detekt configuration and warnings
Add the detekt plugin to the :thumbprint project. Previously it was only be
applied to the root project, and so running it was effectively a no-op since the
root project has no source files itself.

Fix warnings for ClassOrdering which has gotten more strict in 1.18 and disable
the UseDataClass rule since we have a few cases that aren't trivial to migrate
(but probably should be data classes ideally).