Remove unnecessary repositories
We don't need any repositories in the root project, since it doesn't have any
depedencies. We don't need to declare repositories for allprojects{} since we're
declaring them in :thumbprint anyway (which is probably a slightly better place,
to keep them close to the dependencies). Finally, we don't need the JCenter
repository - which is good, since it has been deprecated and will go offline in
~6 months.