Gradle 8.14.3/AGP 8.14.1/Kotlin 2.2.21/Java 17/Min SDK 28/Target SDK 35 (#64)
This upgrades the Thumbprint Android library to the aforementioned
Gradle, AGP, Kotlin and Java versions, and updates the Min SDK and Target SDK
to align with those updates and to be congruent with the SDK properties
we have in the main Thumbtack app. These updates all had to be done
together due to cross-compatibility requirements with all of those
SDKs and frameworks. As part of this upgrade, we also had to update
Robolectric to the latest version of 4.16, and also updated the `sdk`
Robolectric property to 35, to match the new Target SDK.

As part of this upgrade, we also had to:
* Remove Kotlin Synthetics, and convert to View Binding for those classes
that used to use Synthetics.
* Make lint changes with the upgrade to Kotlin 2.2.21.
* Update Robolectric to the latest version of 4.16, and also updated the `sdk`
Robolectric property to 35, to match the new Target SDK.

And finally, we needed to keep Transitive R Classes turned on so that it would
not impact the namespace of the Thumbprint Tokens (colors, spacing, etc) as
they are used in the Thumbtack apps. We can revisit turning this on later.