Fix background for ThumbprintBanner (#66)
When we moved to View Binding, it broke setting the background color of
ThumbprintBanner. In the screenshot tests for our apps, the background got
set on the parent of the banner rather than the root view of the banner
itself. We tried several things, like setting the background on
binding.root and setting it on `this`, but none of those worked.
So to fix this for now, we are wrapping the entire banner in a FrameLayout
to ensure it is isolated from its parent.